--- Engine/Core/Interface/InterfaceDefaultImpl.cpp
@@ -1,6 +1,0 @@
-// InterfaceDefaultImpl.cpp
-// created on 2021/7/16
-// author @zoloypzuo
-// keep interface header compiled
-#include "ZeloPreCompiledHeader.h"
-#include "ISerializable.h"
